% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI_Power_Old.R
\name{ZI_Power}
\alias{ZI_Power}
\title{ZI_Power}
\usage{
ZI_Power(
  model,
  family,
  cov_interest,
  data,
  nsim,
  grid = seq(100, 1000, 100),
  alpha,
  padj = 0
)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{family}{PARAM_DESCRIPTION}

\item{cov_interest}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{nsim}{PARAM_DESCRIPTION}

\item{grid}{PARAM_DESCRIPTION, Default: seq(100, 1000, 100)}

\item{cort}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Performs power analysis using Zero-Inflated count GLM regression model.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pscl]{zeroinfl}}
\code{\link[parameters]{p_value}}
}
