% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI_Power.R
\name{ZI_Power}
\alias{ZI_Power}
\title{ZI_Power}
\usage{
ZI_Power(
  model,
  family,
  cov_interest,
  data,
  nsim,
  grid = seq(100, 1000, 100),
  alpha = 0.05,
  padj = 0
)
}
\arguments{
\item{model}{An R model formula specifying the regression relationship to be used.}

\item{family}{A character value of either “poisson” or “negbin”, which specifies the family that the count distribution in the zero-inflated GLM model should follow.}

\item{cov_interest}{A character string for the name of the predictor you are interested in calculating power for.}

\item{data}{A data frame containing the data set to be used in fitting the initial model.}

\item{nsim}{An integer value indicating the number of simulation iterations to perform for each different sample size value.}

\item{grid}{A vector which specifies the sequence of sample size values to perform the calculation over, Default: seq(100, 1000, 100)}

\item{alpha}{The significance level you would like to use when calculating the proportion of times p-values were less than a certain size, Default: .05}

\item{padj}{Character string specifying a p-value adjustment from the options available for the stats::p.adjust function.  Default: 0}
}
\value{
A list object with two components: Results and Plot. Results is a data.frame object containing the sample size values used and the corresponding power calculation for each sample size. Plot contains a ggplot object which create a power curve showing the power as a function of sample size.
}
\description{
This function calculates power with Monte-Carlo simulation when using zero-inflated count GLM models.
}
\details{
For the padj argument, a value of 0 gives no correction and a string of any method in p.adjust() documentation will perform the corresponding adjustment in each power calculation.

A progress bar was added to show completion percentage in real time. This can be useful since finer grids require high computation time and the progress bar can help estimate how long the calculations will take.
}
\examples{
\dontrun{
Below code corresponds to usage based on the model outlined in our research paper:
 ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),padj = 0)
 ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),padj = "BY")
}
}
\seealso{
\code{\link[pacman]{p_load}}
\code{\link[pscl]{zeroinfl}}
\code{\link[parameters]{p_value}}
}
