% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPowerAnalysis-package.R
\docType{package}
\name{ZIPowerAnalysis-package}
\alias{ZIPowerAnalysis}
\alias{ZIPowerAnalysis-package}
\title{ZIPowerAnalysis: What the Package Does (One Line, Title Case)}
\description{
This package contains functions for the method outlined in Power Analysis for Zero-Inflated Count Models.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
