% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI_Power.R
\name{ZI_Power}
\alias{ZI_Power}
\title{ZI_Power}
\usage{
ZI_Power(
  model,
  family,
  cov_interest,
  data,
  nsim,
  grid = seq(100, 1000, 100),
  alpha,
  padj = 0
)
}
\arguments{
\item{model}{An R model formula specifying the regression relationship.}

\item{family}{A string values of either “poisson” or “negbin”, which specify which family the count distribution in the zero-inflated GLM model should follow.}

\item{cov_interest}{A string for the name of the predictor you are interested in calculating power for.}

\item{data}{A data frame containing the data set to be used in fitting the initial model.}

\item{nsim}{An integer value indicating the number of simulation interactions to perform for each different sample size value}

\item{grid}{A vector which specifies the sequence of sample size values to perform the calculation over, Default: seq(100, 1000, 100)}

\item{alpha}{The significance level you would like to be used when calculating the proportion of times p-values were less than a certain size, Default: .05}

\item{padj}{Specification for a p-value adjustment using the options available for the stats::p.adjust function.  Default: 0}
}
\value{
A list object with two components: Results and Plot. Results is a data.frame object containing the sample size values used and the corresponding power calculation for each sample size. Plot contains a ggplot object which create a power curve showing the power as a function of sample size.
}
\description{
This function calculates power with Monte-Carlo simulation when using zero-inflated count GLM models.
}
\details{
For the padj argument, a value of 0 gives no correction and a string of any method in p.adjust() documentation will perform the corresponding adjustment in each power calculation.
}
\examples{
\dontrun{
if(interactive()){
 #ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),cort = 0)
 #ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),cort = "BY")
 }
}
}
\seealso{
\code{\link[pacman]{p_load}}
\code{\link[pscl]{zeroinfl}}
\code{\link[parameters]{p_value}}
}
