% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI_Power.R
\name{ZI_Power}
\alias{ZI_Power}
\title{ZI_Power}
\usage{
ZI_Power(
  model,
  family,
  cov_interest,
  data,
  nsim,
  grid = seq(100, 1000, 100),
  alpha,
  padj = 0
)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{family}{PARAM_DESCRIPTION}

\item{cov_interest}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{nsim}{PARAM_DESCRIPTION}

\item{grid}{PARAM_DESCRIPTION, Default: seq(100, 1000, 100)}

\item{alpha}{the specified significance level, Default: .05}

\item{padj}{specification for an option p-value adjustment using options in stats::p.adjust documentation. Value of 0 gives no correct and a string of any method in p.adjust() documentation can be used.  Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),cort = 0)
 #ZI_Power(model=Depression~Sex+EOD_total,cov_interest = "EOD_total",family = "poisson",data = dat,nsim = 500,grid = seq(200,4000,100),cort = "BY")
 }
}
}
\references{
Benjamini, Y., and Yekutieli, D. (2001). The control of the false discovery rate in multiple testing under dependency. Annals of Statistics, 29, 1165–1188. doi:10.1214/aos/1013699998.
}
\seealso{
\code{\link[pacman]{p_load}}
\code{\link[pscl]{zeroinfl}}
\code{\link[parameters]{p_value}}
}
